PART 2: TRIGGERS (COMPOSITETRIGGERS)
COMPOSITE TRIGGERS

So the triggers from the previous page are mildly interesting, but we want to do better: we want to 'compose' the earlier triggers, to set up more powerful alert rules. For instance, we may want to raise an alert only when both "google" and "stock" were present in the news item (an idea we can't express right now).

Note that these triggers are not word triggers and should not be subclasses of WordTrigger.

PROBLEM 6

Implement a NOT trigger (NotTrigger).

This trigger should produce its output by inverting the output of another trigger. The NOT trigger should take this other trigger as an argument to its constructor. (Why its constructor? Because we can't change what parameters evaluate takes in... that'd break our polymorphism). So, given a trigger T and a news item x, the output of the NOT trigger's evaluate method should be equivalent to not T.evaluate(x).

When this is done, the NotTrigger unit tests should pass.

Canopy specific instructions:Every time you modify code inps7.pygo to
Run -> Restart Kernel (or hit the CTRL with the dot on your keyboard)
before running ps7_test.py. You have to do this every time you modify the fileps7.pyand want to run the fileps7_test.py, otherwise changes to the former will not be incorporated in the latter.

PROBLEM 7

Implement an AND trigger (AndTrigger).

This trigger should take two triggers as arguments to its constructor, and should fire on a news story only if both of the inputted triggers would fire on that item.

When this is done, the AndTrigger unit tests should pass.

PROBLEM 8

Implement an OR trigger (OrTrigger).

This trigger should take two triggers as arguments to its constructor, and should fire if either one (or both) of its inputted triggers would fire on that item.

When this is done, the OrTrigger unit tests should pass.

Note: In addition to the three classes in this problem (NotTrigger, AndTrigger, and OrTrigger), please provide your definitions of WordTrigger and TitleTrigger in the following box.